% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_clusters.R
\name{c_clusters}
\alias{c_clusters}
\title{c_clusters Function}
\usage{
c_clusters(type, from = 2, to = 50, sd = 0.1, mean = 0, C = c(0.01,
  0.1, 0.5))
}
\arguments{
\item{type}{can be "random", "pred_prey" or "mixture" depending on the type of matrix we want to use for the network}

\item{from}{the smallest number of species in the network (default = 2)}

\item{to}{the biggest number of species in the network (default = 50)}

\item{sd}{the standard deviation used if "random" is chosen (default = 0.1)}

\item{mean}{the mean used if "random" is chosen (default = 0)}

\item{C}{the vector of size 3 containing different  values for the connectance C (default = c(0.01,0.1,0.5))}
}
\description{
This function returns the graph of the number of clusters for different sizes of population
}
