% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_info.R
\name{clusters_info}
\alias{clusters_info}
\title{clusters_info Function}
\usage{
clusters_info(type, S, mean = 0, sd = 0.1, C = 0.1, precision = 0.001)
}
\arguments{
\item{type}{the type of the network}

\item{S}{size of the population}

\item{precision}{the treshold size (default = 0.001)}
}
\description{
This function returns the number of clusters and their average size
}
