% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stable.R
\name{is_stable}
\alias{is_stable}
\title{is_stable Function}
\usage{
is_stable(M)
}
\arguments{
\item{M}{A matrix}
}
\description{
This function returns the stability of a network by calculating its eigenvalues
}
