% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_matrix.R
\name{mixture_matrix}
\alias{mixture_matrix}
\title{mixture_matrix Function}
\usage{
mixture_matrix(S, mean, sd, C)
}
\arguments{
\item{S}{the size of the population}

\item{mean}{the mean used in the gaussian distribution}

\item{sd}{the standard deviation used in the gaussian distribution}

\item{C}{the connectance}
}
\description{
This function returns a matrix corresponding to a network with a mixture of competition and mutualism
}
