% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_frame.R
\name{network_frame}
\alias{network_frame}
\title{network_frame Function}
\usage{
network_frame(type, from = 2, to = 50, mean = 0, sd = 0.1, C = 0.1,
  MC = 1, precision = 0.01)
}
\arguments{
\item{type}{the type of networks we want to create (random, pred_prey, mixture)}

\item{from}{smallest size of population}

\item{to}{biggest size of population}

\item{mean}{the mean used in the gaussian distribution}

\item{sd}{the standard deviation used in the gaussian distribution}

\item{C}{the connectance}

\item{MC}{set it if MC simulations are desired}

\item{precision}{is the treshold size used to convert to binary networks}
}
\description{
This function returns a tidyverse tibble containing networks of a certain type (random/predator-prey/mixture)
and statistics on them: type, network size, average degree, number of clusters, average size of clusters, 
stability of the network, connectance, mean and sd
}
