% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_prey_matrix.R
\name{predator_prey_matrix}
\alias{predator_prey_matrix}
\title{predator_prey_matrix Function}
\usage{
predator_prey_matrix(S, mean, sd, C)
}
\arguments{
\item{S}{the size of the population}

\item{mean}{the mean used in the gaussian distribution}

\item{sd}{the standard deviation used in the gaussian distribution}

\item{C}{the connectance}
}
\description{
This function returns a matrix corresponding to a network with predator/prey interactions
}
