% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_matrix.R
\name{random_matrix}
\alias{random_matrix}
\title{random_matrix Function}
\usage{
random_matrix(S, mean, sd, C)
}
\arguments{
\item{S}{the size of the population}

\item{mean}{the mean used in the gaussian distribution}

\item{sd}{the standard deviation used in the gaussian distribution}

\item{C}{the connectance}
}
\description{
This function returns a random matrix corresponding to a network
}
