% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{relationships}
\alias{relationships}
\title{relationships Function}
\usage{
relationships(type, S, mean = 0, sd = 0.1, C = 0.1, precision = 0.001)
}
\arguments{
\item{type}{the type of the network}

\item{S}{size of the population}

\item{mean}{used in the gaussian law}

\item{sd}{standard deviation used in the gaussian law}

\item{C}{connectance of the network}

\item{precision}{the treshold size}
}
\description{
This function returns the number of edges in the network and 
two list showing every single relationship of the network.~
}
