% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_clusters.R
\name{sd_clusters}
\alias{sd_clusters}
\title{sd_clusters Function}
\usage{
sd_clusters(from = 2, to = 50, sd = c(0.01, 0.25, 0.8), mean = 0,
  C = 0.1)
}
\arguments{
\item{from}{the smallest number of species in the network (default = 2)}

\item{to}{the biggest number of species in the network (default = 50)}

\item{sd}{the standard deviation used (default = 0.1)}

\item{mean}{the mean used if "random" is chosen (default = c(0.01,0.25,0.8))}

\item{C}{the connectance (default = 0.1)}
}
\description{
This function returns the graph of the size of clusters for different values of standard deviation (in the case of a random matrix)
}
